ZOL-shop/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==1.11.11
django-redis==4.10.0
get==2019.4.13
idna==2.8
Pillow==6.1.0
post==2019.4.13
public==2019.4.13
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.1
pytz==2019.2
query-string==2019.4.13
redis==3.3.8
requests==2.22.0
urllib3==1.25.5
